move component to examples and add to workspace
