+++ Cargo.toml
@@ -38,6 +38,7 @@ members = [
     "crates/fluvio-test-derive",
     "crates/fluvio-test-util",
     "crates/fluvio-types",
+    "examples/component",
 ]
 resolver = "2"
