+++ examples/component/Cargo.toml
@@ -11,4 +11,4 @@ crate-type = ['cdylib']
 [dependencies]
 regex = "1.5"
 once_cell = "1.13.0"
-fluvio-smartmodule = { path = "../../" }
+fluvio-smartmodule = { path = "../../crates/fluvio-smartmodule" }
